NAT: 64-bit byte counter.
git-svn-id: https://www.virtualbox.org/svn/vbox/trunk@42010 cfe28804-0f27-0410-a406-dd0f0b0b656f

+++ src/VBox/Devices/Network/slirp/tftp.c
@@ -67,7 +67,7 @@ typedef struct TFTPSESSION
     struct      in_addr IpClientAddress;
     uint16_t    u16ClientPort;
     int         iTimestamp;
-    int         cbTransfered;
+    uint64_t    cbTransfered;
     ENMTFTPSESSIONFMT enmTftpFmt;
     TFPTPSESSIONOPTDESC OptionBlkSize;
     TFPTPSESSIONOPTDESC OptionTSize;
